Fix timestamp types

+++ firmware/dwm1001_dev_source_code/development/dedicated_time_protocol/modules/message_template/message_template.h
@@ -45,12 +45,12 @@ typedef struct {
 } msg_template;
 
 /* Function prototypes */
-void initTable(uint32 table[NUM_STAMPS_PER_CYCLE][N]);
+void initTable(uint64 table[NUM_STAMPS_PER_CYCLE][N]);
 void convertToStruct(uint8 *array, msg_template *msg);
 void convertToArr(msg_template msg, uint8 *array);
-void updateTable(uint32 table[NUM_STAMPS_PER_CYCLE][N], msg_template msg, uint32 ts, uint8 thisId);
-void getFullTs(uint32 table[NUM_STAMPS_PER_CYCLE][N], uint32 ts[NUM_STAMPS_PER_CYCLE], uint8 thisId, uint8 targetId);
-void getHalfTs(uint32 table[NUM_STAMPS_PER_CYCLE][N], uint32 ts[NUM_STAMPS_PER_CYCLE/2], uint8 thisId, uint8 targetId);
+void updateTable(uint64 table[NUM_STAMPS_PER_CYCLE][N], msg_template msg, uint64 ts, uint8 thisId);
+void getFullTs(uint64 table[NUM_STAMPS_PER_CYCLE][N], uint64 ts[NUM_STAMPS_PER_CYCLE], uint8 thisId, uint8 targetId);
+void getHalfTs(uint64 table[NUM_STAMPS_PER_CYCLE][N], uint64 ts[NUM_STAMPS_PER_CYCLE/2], uint8 thisId, uint8 targetId);
 
 /* Public variables */
 extern uint64 tsTable[NUM_STAMPS_PER_CYCLE][N];
